# api/openapi.yaml
openapi: 3.0.3
info:
  title: AgentGuard Policy Engine API
  description: |
    **AgentGuard** is a runtime security platform for AI agents. It provides policy-based evaluation,
    audit trails, PII detection, MCP server governance, agent hierarchy management, and OWASP compliance reporting.

    ## Authentication

    AgentGuard uses three authentication schemes:

    - **API Key** (`x-api-key` header): Your tenant API key, prefixed `ag_live_`. Used for most authenticated endpoints.
    - **Agent Key** (`x-api-key` header): Agent-scoped key, prefixed `ag_agent_`. Used for evaluate calls from specific agents.
    - **Admin Key** (`x-admin-key` header or `x-api-key`): Platform admin key. Required for admin-only endpoints.

    ## Rate Limits

    - Unauthenticated: 10 requests/min per IP
    - Authenticated: 100 requests/min per IP
    - Signup: 5 per hour per IP

    ## Base URL

    Production: `https://api.agentguard.tech/api/v1`
  version: 0.10.0
  contact:
    name: AgentGuard Support
    url: https://agentguard.tech
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.agentguard.tech/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Tenant API key (prefixed `ag_live_`) or agent key (prefixed `ag_agent_`)
    AdminKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Platform admin key (set via ADMIN_KEY environment variable)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token (future use)
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Invalid API key"
        details:
          type: array
          items:
            type: string
          description: Additional error details (validation errors)
...
paths:
...
  /alerts:
    get:
      tags: [Alerts]
      summary: List alerts
      description: |
        List anomaly-detection alerts for the authenticated tenant.
        Optionally filter by severity or resolved state.
      operationId: listAlerts
      security:
        - ApiKeyAuth: []
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warning, critical]
          description: Filter by severity level
        - name: resolved
          in: query
          schema:
            type: boolean
          description: "Filter by resolved state (true=resolved, false=open)"
      responses:
        '200':
          description: Alert list
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlertObject'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /alerts/{id}/acknowledge:
    post:
      tags: [Alerts]
      summary: Acknowledge alert
      description: Mark an alert as acknowledged (resolved). Idempotent for already-resolved alerts.
      operationId: acknowledgeAlert
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Alert ID
      responses:
        '200':
          description: Alert acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  acknowledged:
                    type: boolean
        '400':
          description: Alert already resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Alert not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /alerts/rules:
    get:
      tags: [Alerts]
      summary: List anomaly rules
      description: List custom anomaly detection rules for the authenticated tenant.
      operationId: listAnomalyRules
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Anomaly rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnomalyRule'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Alerts]
      summary: Create anomaly rule
      description: Create a custom anomaly detection rule that generates alerts when triggered.
      operationId: createAnomalyRule
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, metric, condition, threshold, windowMinutes]
              properties:
                name:
                  type: string
                  maxLength: 200
                  description: Human-readable rule name
                  example: "High block rate alert"
                metric:
                  type: string
                  enum: [block_rate, evaluate_volume, unique_tools, error_rate, latency_p99]
                  description: Metric to monitor
                condition:
                  type: string
                  enum: [gt, lt, spike, drop]
                  description: Comparison operator
                threshold:
                  type: number
                  description: Threshold value to compare against
                  example: 0.2
                windowMinutes:
                  type: integer
                  minimum: 1
                  maximum: 1440
                  description: Evaluation window in minutes
                  example: 60
                severity:
                  type: string
                  enum: [info, warning, critical]
                  default: warning
                enabled:
                  type: boolean
                  default: true
            example:
              name: "High block rate"
              metric: "block_rate"
              condition: "gt"
              threshold: 0.2
              windowMinutes: 60
              severity: "warning"
      responses:
        '201':
          description: Rule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnomalyRule'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /alerts/rules/{id}:
    put:
      tags: [Alerts]
      summary: Update anomaly rule
      description: Update an existing anomaly detection rule. All fields are optional (partial update).
      operationId: updateAnomalyRule
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Rule ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 200
                metric:
                  type: string
                  enum: [block_rate, evaluate_volume, unique_tools, error_rate, latency_p99]
                condition:
                  type: string
                  enum: [gt, lt, spike, drop]
                threshold:
                  type: number
                windowMinutes:
                  type: integer
                  minimum: 1
                  maximum: 1440
                severity:
                  type: string
                  enum: [info, warning, critical]
                enabled:
                  type: boolean
      responses:
        '200':
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnomalyRule'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Alerts]
      summary: Delete anomaly rule
      description: Delete a custom anomaly detection rule.
      operationId: deleteAnomalyRule
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Rule ID
      responses:
        '200':
          description: Rule deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  deleted:
                    type: boolean
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── AUDIT (verify + export) ───────────────────────────────────────────────
...